Updated DataScript folder
Renamed all the script files.
Added comments to all datascripts files.
